// src/modules/general/enums/labelHome.ts
export enum LabelHome{
    seccionPrimeraTitulo = 'FIREPLOY',
    seccionPrimeraCuerpo = 'Despliega y muéstrate cómo todo un DevOps, desde el alcance de tu mano',
    seccionPrimeraBoton = 'Comienza ahora',
    seccionSegundaTitulo = 'Explora Proyectos',
    seccionSegundaCuerpo = 'Te permitimos visualizar los demás proyectos desplegados con la plataforma, para que te inspires y aprendas de los demás',
    seccionSegundaBoton = 'Explora proyectos',
    seccionTerceraTitulo = 'Explora Portafolios',
    seccionTerceraCuerpo = 'Te permitimos compartir tus logros y ver los de los demás. Conoce gente y desarrolla con ellos',
    seccionTerceraBoton = 'Explora Portafolios'
}